+++ blocklets/ai-studio/src/components/file-editor/use-call-agent-output.tsx
@@ -1,7 +1,7 @@
 import { useProjectStore } from '@app/pages/project/yjs-state';
 import { useLocaleContext } from '@arcblock/ux/lib/Locale/context';
 import { AssistantYjs } from '@blocklet/ai-runtime/types';
-import { cloneDeep, cloneDeepWith, sortBy } from 'lodash';
+import { cloneDeep, sortBy } from 'lodash';
 import { useCallback, useMemo } from 'react';
 
 import { getOutputName, runtimeOutputVariables } from './output/type';
@@ -19,31 +19,31 @@ const useCallAgentOutput = ({
   const { getFileById } = useProjectStore(projectId, gitRef);
 
   const id = assistant.type === 'callAgent' ? assistant.call?.id : undefined;
+
   const getOutputs = useCallback(() => {
-    if (assistant.type === 'callAgent' && assistant.call) {
-      const callAgent = getFileById(assistant.call.id);
-      return (callAgent?.outputVariables && sortBy(Object.values(callAgent.outputVariables), 'index')) || [];
-    }
+    if (assistant.type !== 'callAgent' || !assistant.call) return [];
 
-    if (assistant.type === 'parallelCallAgent') {
-      const agents = Object.values(assistant?.agents || {}).map((i) => i.data);
-      return agents.flatMap((i) => {
-        const callAgent = getFileById(i.id);
-        return (callAgent?.outputVariables && sortBy(Object.values(callAgent.outputVariables), 'index')) || [];
-      });
-    }
-
-    return [];
+    const callAgent = getFileById(assistant.call.id);
+    return (callAgent?.outputVariables && sortBy(Object.values(callAgent.outputVariables), 'index')) || [];
+    // const agents = Object.values(assistant?.agents || {}).map((i) => i.data);
+    // return agents.flatMap((i) => {
+    //   const callAgent = getFileById(i.id);
+    //   return (callAgent?.outputVariables && sortBy(Object.values(callAgent.outputVariables), 'index')) || [];
+    // });
   }, [id]);
 
-  const appearance = runtimeOutputVariables.find((i) => i.group === 'appearance')?.outputs || [];
-  const outputs = useMemo(() => {
-    const list = getOutputs().map((item) => item.data);
-    const map = new Map();
-    list.forEach((i) => map.set(i.name, i)); // 去重
+  const appearanceOutputs = useMemo(
+    () => runtimeOutputVariables.find((i) => i.group === 'appearance')?.outputs || [],
+    []
+  );
 
-    return Array.from(map.values()).filter((i) => !appearance.find((r) => r.name === i.name)); // 过滤外观输出变量
-  }, [getOutputs]);
+  const outputs = useMemo(
+    () =>
+      getOutputs()
+        .map((item) => item.data)
+        .filter((i) => !appearanceOutputs.find((r) => r.name === i.name)), // 过滤外观输出变量
+    [getOutputs]
+  );
 
   const outputsExist = useMemo(
     () =>