+++ blocklets/ai-runtime/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.3.17 (2024-7-3)
+
+- fix: correct parsing multiple function calling chunks
+
+## 0.3.16 (2024-7-3)
+
+- chore: export aigne components in aigne-sdk
+- chore: add api to query all agents from aigne studio
+- fix: menus do not appear when clicking the menu button on examples
+
+## 0.3.15 (2024-7-2)
+
+- chore: disable submit for the invalid inputs
+- chore: support multiline text for question input
+- fix: correct og meta for agent previewer
+
 ## 0.3.14 (2024-7-1)
 
 - fix: only png files are allowed for the project icon